manager/requirements.txt
google-cloud-storage
python-dotenv
kubernetes
PyYAML
